requirements.txt
flask
flask-cors
google-cloud-documentai
python-dotenv
proto-plus
pypdf
openai

# Rule-based extraction dependencies
pandas
fuzzywuzzy[speedup]
python-Levenshtein
regex
